Add the LOGIN command
This change introduces the LOGIN command, which allows the user to log into
the EndBASIC cloud service.

On its own, this change does not let the user achieve much, given that all
that's included here is the ability to log into the server.  However, this
commit has to add a lot of boilerplate, specifically around testing, so it
belongs in its own commit.

The implementation of the cloud-backed drive, as well as further tweaks to
the login command, will come later.