Run the cloud service tests on GitHub Actions
Include the ignored tests in our test runs so that we validate the cloud
service client on all platforms.  These tests require specific environment
variables to be set, which are already configured in the GitHub Actions.

I'm not sure this will be a good idea in the long run given that these
tests will likely turn flaky at some point... but given that I'm actively
iterating on this right now, it seems to make sense to include them.