Add a script to run the service integration tests
The new Makefile is intended to only have a simple "make test" target
that runs all ignored tests after sourcing a local-only configuration
file that includes the required account details.

I'd rather use the [env] section of Cargo's configuration to achieve
this, but unfortunately that's only a nightly feature for now.